CDM-20 Refactor with correct model and mapper package location
