+++ demo-app/src/main/java/com/example/demo/controller/AddressController.java
@@ -1,7 +1,7 @@
 package com.example.demo.controller;
 
-import com.example.demo.controller.mappers.CustomerAddressMapper;
-import com.example.demo.controller.mappers.CustomerMapper;
+import com.example.demo.mappers.CustomerAddressMapper;
+import com.example.demo.mappers.CustomerMapper;
 import com.example.demo.models.CustomerAddress;
 import com.example.demo.models.CustomerResponse;
 import com.example.demo.service.CustomerService;
