+++ demo-app/src/main/java/com/example/demo/models/CustomerAddressDTO.java
@@ -1,4 +1,4 @@
-package com.example.demo.controller.models;
+package com.example.demo.models;
 
 import lombok.Builder;
 import lombok.Data;
