+++ demo-app/src/main/java/com/example/demo/models/entities/Customer.java
@@ -1,4 +1,4 @@
-package com.example.demo.controller.models.entities;
+package com.example.demo.models.entities;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
