+++ demo-app/src/main/java/com/example/demo/service/CustomerService.java
@@ -1,10 +1,10 @@
 package com.example.demo.service;
 
-import com.example.demo.controller.mappers.CustomerAddressMapper;
-import com.example.demo.controller.mappers.CustomerMapper;
-import com.example.demo.controller.models.CustomerAddressDTO;
-import com.example.demo.controller.models.CustomerDTO;
-import com.example.demo.controller.models.entities.Customer;
+import com.example.demo.mappers.CustomerAddressMapper;
+import com.example.demo.mappers.CustomerMapper;
+import com.example.demo.models.CustomerAddressDTO;
+import com.example.demo.models.CustomerDTO;
+import com.example.demo.models.entities.Customer;
 import com.example.demo.repository.CustomerRepository;
 import jakarta.annotation.Nullable;
 import jakarta.validation.Valid;
