+++ demo-app/src/test/java/com/example/demo/controller/mappers/CustomerAddressMapperTest.java
@@ -1,7 +1,8 @@
 package com.example.demo.controller.mappers;
 
-import com.example.demo.controller.models.CustomerAddressDTO;
-import com.example.demo.controller.models.entities.Address;
+import com.example.demo.models.CustomerAddressDTO;
+import com.example.demo.models.entities.Address;
+import com.example.demo.mappers.CustomerAddressMapper;
 import org.junit.jupiter.api.Test;
  import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
